simple_service/app/webui/requirements.txt
streamlit==1.36.0
requests==2.32.3
extra-streamlit-components==0.1.71
fastapi==0.111.0
pydantic==2.3.0
pydantic-settings==2.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
uvicorn==0.30.1
sqlmodel==0.0.19
pika==1.2.0
joblib==1.4.2
passlib==1.7.4
python-jose==3.3.0